requirements.txt
absl-py==0.7.1
alabaster==0.7.12
astor==0.7.1
awscli==1.16.137
Babel==2.6.0
botocore==1.12.127
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.3.9
coverage==4.5.3
cycler==0.10.0
Cython==0.29.6
darkflow==1.0.0
docutils==0.14
entrypoints==0.3
flake8==3.7.7
gast==0.2.2
google-streetview==1.2.9
grpcio==1.19.0
h5py==2.9.0
idna==2.8
imagesize==1.1.0
Jinja2==2.10
jmespath==0.9.4
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
kwconfig==1.1.7
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mock==2.0.0
numpy==1.16.2
opencv-python==4.0.0.21
packaging==19.0
pbr==5.1.3
Pillow==6.0.0
protobuf==3.7.1
pyasn1==0.4.5
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.3.1
pyparsing==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.9
PyYAML==3.13
requests==2.21.0
rsa==3.4.2
s3transfer==0.2.0
scipy==1.2.1
six==1.12.0
snowballstemmer==1.2.1
Sphinx==2.0.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.15.2
